move validator to separate file
