+++ src/members/models/member.py
@@ -7,21 +7,7 @@ from django.utils import timezone
 from django.utils.translation import ugettext_lazy as _
 from simple_email_confirmation.models import SimpleEmailConfirmationUserMixin
 from utils.melos_client import MelosClient
-
-
-class SSNValidator(validators.RegexValidator):
-    def __init__(self):
-        super(SSNValidator, self).__init__(
-            # This regex makes sure that the year is either 1900 or 2000
-            regex=r'^[1-2][0|9][0-9]{2}[0-1][0-9][0-3][0-9][-][0-9]{4}$|'  # noqa: E501, YYYYMMDD-XXXX
-                  r'^[1-2][0|9][0-9]{2}[0-1][0-9][0-3][0-9][0-9]{4}$|'  # noqa: E501, YYYYMMDDXXXX
-                  r'^[0|9][0-9]{1}[0-1][0-9][0-3][0-9][-][0-9]{4}$|'  # noqa: E501, YYMMDD-XXXX
-                  r'^[0|9][0-9]{1}[0-1][0-9][0-3][0-9][0-9]{4}$',  # noqa: E501, YYMMDDXXXX
-            message=_(
-                'Use the format YYYYMMDD-XXXX, YYMMDD-XXXX, \
-                YYYYMMDDXXXX, YYMMDDXXXX for your ssn.'
-            )
-        )
+from utils.validators import SSNValidator
 
 
 class CaseInsensitiveUsernameUserManager(UserManager):
@@ -157,7 +143,8 @@ class Member(SimpleEmailConfirmationUserMixin, AbstractUser):
     melos_id = models.IntegerField(
         blank=True,
         editable=False,
-        null=True
+        null=True,
+        unique=True,
     )
 
     melos_user_data = None
@@ -286,6 +273,7 @@ class Member(SimpleEmailConfirmationUserMixin, AbstractUser):
     @staticmethod
     def find_by_ssn(ssn):
         try:
+            ssn = ssn.strip()
             SSNValidator()(ssn)
             melos_id = MelosClient.get_melos_id(ssn)
             return Member.find_by_melos_id(melos_id)
